Collapse options into main file
