+++ manifestival.go
@@ -39,6 +39,23 @@ type Manifest struct {
 
 var _ Manifestival = &Manifest{}
 
+// Option follows the "functional object" idiom
+type Option func(*Manifest)
+
+// UseLogger will cause manifestival to log its actions
+func UseLogger(log logr.Logger) Option {
+	return func(m *Manifest) {
+		m.log = log
+	}
+}
+
+// UseClient enables interaction with the k8s API server
+func UseClient(client Client) Option {
+	return func(m *Manifest) {
+		m.Client = client
+	}
+}
+
 // NewManifest creates a Manifest from a comma-separated set of YAML
 // files, directories, or URLs. It's equivalent to
 // `ManifestFrom(Path(pathname))`
